+++ src/pages/tu/index.ts
@@ -1,3 +1,3 @@
 export { MyCoursesPage, MyContentPage, CourseCreatePage, TuContentCreatePage, ContentDetailPage } from './teaching';
 export { SettingsLanguagePage } from './settings';
-export { LandingPage } from './LandingPage';
+export { LandingPage, Page1, Page2, Page3 } from './main';
