+++ src/pages/tu/main/Page1.tsx
@@ -1,0 +1,10 @@
+export function Page1() {
+  return (
+    <div className="flex items-center justify-center min-h-[60vh]">
+      <div className="text-center space-y-4">
+        <h1 className="text-2xl font-bold text-text-primary">1페이지</h1>
+        <p className="text-text-secondary">개발 예정입니다.</p>
+      </div>
+    </div>
+  );
+}
