+++ src/pages/tu/main/index.ts
@@ -1,0 +1,4 @@
+export { LandingPage } from './LandingPage';
+export { Page1 } from './Page1';
+export { Page2 } from './Page2';
+export { Page3 } from './Page3';
